requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0

# AI/ML
google-generativeai>=0.3.0
openai>=1.0.0  # Optional: for GPT integration

# Data Processing
openpyxl>=3.1.0  # Excel support
xlsxwriter>=3.1.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
altair>=5.0.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0
requests>=2.31.0
validators>=0.22.0

# UI/UX Enhancements
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4
streamlit-modal>=0.1.0
streamlit-elements>=0.1.0

# Performance & Caching
redis>=5.0.0  # Optional: for advanced caching
cachetools>=5.3.0

# Security
cryptography>=41.0.0
bcrypt>=4.0.0

# Development (Optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Monitoring & Logging
loguru>=0.7.0
psutil>=5.9.0  # System monitoring

# Export Capabilities
fpdf2>=2.7.0  # PDF generation
reportlab>=4.0.0
Pillow>=10.0.0  # Image processing